+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # @project-version@ (@build-time@)
 
 * Fixed a Lua error with Auras widget the occurred when a Priest uses Mindcontrol on another player [GH-520].
+* Improved visual representation of interact icon for target units.